[feature]: implement autoplaynextpodcast, cleanup files
