+++ src/app/dashboard/components/PodSearch.tsx
@@ -3,7 +3,6 @@ import Paper from "@mui/material/Paper";
 import InputBase from "@mui/material/InputBase";
 import Divider from "@mui/material/Divider";
 import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
 import SearchIcon from "@mui/icons-material/Search";
 import PodcastsIcon from "@mui/icons-material/Podcasts";
