+++ src/app/dashboard/page.tsx
@@ -1,51 +1,24 @@
 "use client";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import "../globals.css";
-import { signOutUser } from "../utils/signOut";
 import PodSearch from "./components/PodSearch";
 import { Button } from "@mui/material";
-import AudioPlayer from "./components/AudioPlayer";
 import { useQuery } from "react-query";
-import { getAllPodcasts } from "./components/api";
+import { getAllPodcasts, Podcast } from "../utils/api";
 import { handleApiError } from "../service/axios";
 import CallMadeIcon from "@mui/icons-material/CallMade";
 import PodCard from "./components/PodCard";
 import CircularProgress from "@mui/material/CircularProgress";
 import Box from "@mui/material/Box";
 
-export interface Podcast {
-  _id: string;
-  name: string;
-  slug: string;
-  url: string;
-  user: string;
-  duration: number;
-  createdAt: string;
-  updatedAt: string;
-}
-
-interface Track {
-  url: string;
-  title: string;
-  tags: string[];
-}
-
 export default function Dashboard() {
-  const [audioSrc, setAudioSrc] = useState("");
-  const [error, setError] = useState<string>("");
-  const [showPlayer, setShowPlayer] = useState(false);
-
-  const [tracks, setTracks] = useState<Track[]>([]);
-
   const [podcasts, setPodcasts] = useState<Podcast[]>([]);
   const [currentIndex, setCurrentIndex] = useState<number | null>(null);
 
   const getPodcasts = async () => {
     try {
       const { data } = await getAllPodcasts();
-      console.log("program data", data.data[0]);
-
-      setPodcasts(data.data);
+      setPodcasts(data);
     } catch (error) {
       throw handleApiError(error);
     }
